refactor: splitting out code openapi->jsonschema work into its own library
