+++ .eslintignore
@@ -1,4 +1,3 @@
-__tests__/__fixtures__/
 __tests__/cli/__fixtures__/
 coverage/
 dist/
